refactor editorControls
